make component for field
